Export useDevicePermissionStatus from library
