+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added optional img to MessageAttachment
 - Added a classname to PopOverMenu component for styling access
 - Added forwardRef for Textarea
+- Added `useDevicePermissionStatus` hook as an exported component from the library.
 
 ### Changed
