--- demo/meeting/src/enums/index.ts
@@ -1,9 +1,0 @@
-// Copyright 2020-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-export enum DevicePermissionStatus {
-  UNSET = 'UNSET',
-  IN_PROGRESS = 'IN_PROGRESS',
-  GRANTED = 'GRANTED',
-  DENIED = 'DENIED',
-}
