+++ src/hooks/sdk/useDevicePermissionStatus.tsx
@@ -1,12 +1,12 @@
-// Copyright 2020-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 import { useEffect, useState } from 'react';
-import { useMeetingManager } from 'amazon-chime-sdk-component-library-react';
 
-import { DevicePermissionStatus } from '../enums';
+import { useMeetingManager } from '../../providers/MeetingProvider';
+import { DevicePermissionStatus } from '../../types';
 
-export default function useDevicePermissionStatus() {
+export function useDevicePermissionStatus() {
   const meetingManager = useMeetingManager();
   const [permission, setPermission] = useState<string>(
     DevicePermissionStatus.UNSET
@@ -24,3 +24,5 @@ export default function useDevicePermissionStatus() {
 
   return permission;
 }
+
+export default useDevicePermissionStatus;
