+++ src/index.ts
@@ -136,6 +136,7 @@ export { useMeetingStatus } from './hooks/sdk/useMeetingStatus';
 export { useLocalAudioInputActivity } from './hooks/sdk/useLocalAudioInputActivity';
 export { useLocalAudioInputActivityPreview } from './hooks/sdk/useLocalAudioInputActivityPreview';
 export { useBandwidthMetrics } from './hooks/sdk/useBandwidthMetrics';
+export { useDevicePermissionStatus } from './hooks/sdk/useDevicePermissionStatus';
 
 // Providers
 export { NotificationProvider } from './providers/NotificationProvider';
@@ -160,7 +161,7 @@ export { lightTheme, darkTheme, GlobalStyles, StyledReset } from './theme';
 export { VideoQuality } from './hooks/sdk/useSelectVideoQuality';
 
 // enums
-export { MeetingStatus } from './types/index';
+export { MeetingStatus, DevicePermissionStatus } from './types/index';
 export { Severity, ActionType } from './providers/NotificationProvider';
 
 // Class
