+++ src/types/index.ts
@@ -53,3 +53,10 @@ export type RosterAttendeeType = {
 export type RosterType = {
   [attendeeId: string]: RosterAttendeeType;
 };
+
+export enum DevicePermissionStatus {
+  UNSET = 'UNSET',
+  IN_PROGRESS = 'IN_PROGRESS',
+  GRANTED = 'GRANTED',
+  DENIED = 'DENIED',
+};
